Additional features
- `-vnc` flag will generate a random port that a VM will start on
- `-d` will remove the storage on `vm stop`
- `id` is now checked
- Fixes to error reporting
